requirements.txt
openai>=1.14.0
pydantic>=2.0.0
python-dotenv>=1.0.0
chromadb>=0.4.0
rich>=13.0.0
